Cleaned up the code on the Dj Signup page. Changed ImageContainer component to UploadImage component. This way we can use UploadImage component for the add Event page for the event image upload feature there. I also created two Helper functions (selectImage and uploadImage) and stored them in the utils folder in the Helper.js file. This way they too can be used by the upload event image feature. Eliminates repeat code. I also moved the Cloudinary API to the utils API folder. This way all of our API calls will be in the same folder.
